Move all payment method object creation to custom hook
